# docs/swagger.json
{
  "basePath": "/api",
  "definitions": {
    "models.AuthResponse": {
      "example": {
        "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiZXhwIjoxNzA1MzQzODAwfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
        "user": {
          "created_at": "2024-01-15T10:30:00Z",
          "email": "user@example.com",
          "external_id": "",
          "id": 1,
          "provider": "local",
          "role": "user"
        }
      },
      "properties": {
        "token": {
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiZXhwIjoxNzA1MzQzODAwfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/models.User"
        }
      },
      "type": "object"
    },
    "models.LoginRequest": {
      "properties": {
        "email": {
          "example": "user@example.com",
          "type": "string"
        },
        "password": {
          "example": "password123",
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ],
      "type": "object"
    },
...
    "models.User": {
      "properties": {
        "created_at": {
          "example": "2024-01-15T10:30:00Z",
          "type": "string"
        },
        "email": {
          "example": "user@example.com",
          "type": "string"
        },
        "external_id": {
          "description": "Пустая строка для локальных пользователей",
          "example": "",
          "type": "string"
        },
        "id": {
          "example": 1,
          "type": "integer"
        },
        "provider": {
          "description": "'local', 'google', 'yandex'",
          "example": "local",
          "type": "string"
        },
        "role": {
          "description": "'user', 'admin'",
          "example": "user",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "host": "localhost:8080",
  "info": {
    "contact": {},
    "description": "API for music streaming service.",
    "title": "Koteyye Music API",
    "version": "1.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Login credentials",
            "in": "body",
            "name": "input",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.LoginRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "schema": {
              "$ref": "#/definitions/models.AuthResponse"
            }
          },
          "400": {
            "description": "Bad request - invalid input",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized - invalid credentials",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "User Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/{provider}/callback": {
      "get": {
        "parameters": [
          {
            "description": "OAuth Provider",
            "enum": [
              "google",
              "yandex"
            ],
            "example": "google",
            "in": "path",
            "name": "provider",
            "required": true,
            "type": "string"
          },
          {
            "description": "Authorization Code",
            "example": "4/0AX4XfWhi_abc123xyz",
            "in": "query",
            "name": "code",
            "required": true,
            "type": "string"
          },
          {
            "description": "State Parameter",
            "example": "random_state_string",
            "in": "query",
            "name": "state",
            "type": "string"
          }
        ],
        "responses": {
          "307": {
            "description": "Temporary Redirect to frontend with JWT token in query"
          },
          "400": {
            "description": "Bad request - missing or invalid code",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "OAuth Callback",
        "tags": [
          "oauth"
        ]
      }
    },
    "/auth/{provider}/login": {
      "get": {
        "parameters": [
          {
            "description": "OAuth Provider",
            "enum": [
              "google",
              "yandex"
            ],
            "example": "google",
            "in": "path",
            "name": "provider",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "307": {
            "description": "Temporary Redirect to OAuth provider"
          },
          "400": {
            "description": "Bad request - unsupported provider",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "OAuth Login Initiation",
        "tags": [
          "oauth"
        ]
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}